updated DAO

+++ src/main/java/ua/java/conferences/dao/EntityDAO.java
@@ -8,7 +8,7 @@ public interface EntityDAO<T> {
 
     boolean add(T t) throws DAOException;
 
-    T getById(int id) throws DAOException;
+    T getById(long id) throws DAOException;
 
     List<T> getAll() throws DAOException;
